requirements.txt
setuptools==75.8.2
wheel==0.45.1
python-telegram-bot==20.3